Feature update, shell rework, backend
The SQLite alternative Back-end is now functional however needs more testing. It is still reliable enough that I have not encountered any problems, however it needs to be used according to a narrow specification. Error handling is not implemented properly in the shell yet, however this is trivial to add later.